Added Enemy Roaming/Wandering
Enemies now can free roam around the map. I also adjusted the enemy stats so It feels a bit more manageable of fights, the poison on the spiders move a lot slower now instead of instant hits basically. I also made the zombie movement speed a little bit faster so it is harder for the player to deal with them now.