Started Boss Controller script
I just started with a foundation where if a key is pressed the boss will take damage for now. I used debug logs to identify the stages of the boss. stages are set to a health amount on the boss. when it reaches the stage will be triggered and applied onto the boss. each stage will be set with different attack types and level of difficulty for the player.