Updating Boss Script
I added a movement system for the boss, I decided to not go with a nav mesh agent for now, just a basic tracking player pos within a range. The boss will now do its first attack: dark orb (still work in progress) but it a chaser that will follow the player and deal damage on impact. Player is now also able to deal damage to the boss!