Compass Bar Edit and Compass Markers Fix
Compass Markers fully function and generate on scene load. Objects delete from the list on pick-up and markers disappear from the compass as well.

Added a ParentListener script to help facilitate actions going of the destruction of other options (ie OnDestroy) so you can unlink or run scripts to happen before they are lost to the destroy process.