Artifact functionality up and win condition fixed
Artifacts are required for each scene to be able to progress through the portal. This can be bypassed by ticking the "collected" box for the scene's instance of the end portal.

Dialog now also informs the player that they are missing something to proceed (alluding to the artifact's presence in the level)